Who would believe it? Another 64-bit problem... according to John Daniels the edf format "unsignedlong" is a 32 bit type. Now it opens ID15 summed frelon images...
git-svn-id: https://fable.svn.sourceforge.net/svnroot/fable/fabio/trunk@4697 d066a591-ed20-0410-833c-cb8bbb91b4a7

+++ src/edfimage.py
@@ -24,8 +24,8 @@ DATA_TYPES = {  "SignedByte"     :  np.int8,
                 "UnsignedShortInteger" : np.uint16,
                 "SignedInteger"  :  np.int32,
                 "UnsignedInteger":  np.uint32,
-                "SignedLong"     :  np.int,
-                "UnsignedLong"   :  np.uint,
+                "SignedLong"     :  np.int32,
+                "UnsignedLong"   :  np.uint32,
                 "FloatValue"     :  np.float32,
                 "FLOAT"          :  np.float32, # fit2d
                 "DoubleValue"    :  np.float
